move Win32 SearchPath and SHGetFolderPath into the Win32 package
That completes the transition to Unicode FilePath support in the
directory package on Win32 (#3300).

The main cleanup remaining to do here is to use native Win32
implementations of getPermissions/setPermissions/copyPermissions
instead of calls to stat().

We are very close to being able to remove the last bits of C and
autoconf here...