Fix `createDirectoryIfMissing` silently failing
In some cases, `createDirectoryIfMissing` would silently fail. For
example the following invocation would fail to report via an exception
that it couldn't create a folder:

  let testdir = "/tmp/sometestdir"
  writeFile testdir ""
  createDirectoryIfMissing False testdir

A related issue was the failure to create a folder hierarchy due to lack
of permissions, for instance

  createDirectoryIfMissing True "/foo"

for a non-priviledged user would silently fail (i.e. no exception
thrown), even though "/foo" was not created.

Fixes #4 (see also #10 for discussion)