Add script to run the GHC test framework
The script automatically clones the GHC repo and invokes the test
framework with the correct arguments.  A patch for the GHC repo is also
included to work around bugs/incompatbilities.

Note that it requires an existing Cabal project to work, due to the need
for Cabal-specific files and configuration.  Coupling the tests to Cabal
allows them to be run using the version of project that is in the
current working tree rather than whatever that's installed.  Note that
the project must still be built beforehand.  Also, if a sandbox is used,
the script must be run under `cabal exec`.