Hook test suite to `cabal test`
`cabal test` now invokes the GHC test framework to run the tests.  In
effect, it's just a wrapper that runs `tools/run-tests` with the correct
environment.

This is somewhat complicated by the fact that we cannot add `process` as
a dependency, so we have to reinvent the wheel just to launch the shell
script.