Update test scripts to save local copy of test framework
Nix requires everything to be built from source without external inputs,
so we need to cache our own copy of the GHC test framework.  We could
tar.gz it, but:

  - it's binary, making it hard to compare changes without extracting;
  - metadata such as timestamps can introduce spurious changes.

On the other hand, embedding it as a directory leads to a lot of extra
entries in the `.cabal` file that'd have to be maintained separately.

Therefore we choose to store the test framework in a shell archive.