Remove dependence on the C bits entirely and fix realpath bug
The directory package itself no longer depends on the C bits
(cbits/directory.c, include/HsDirectory.h), but they will remain for
backward compatibility.

This also fixes the potential buffer overflow that can occur in realpath
when PATH_MAX is absent.  Newer versions of POSIX support realpath with
NULL as the second argument, in which case the result is dynamically
allocated by the OS and may be freed using 'free'.  (The hope is that
older systems that don't support this feature would be kind enough to
provide a PATH_MAX, at least.)