Ignore --with-gcc if it contains spaces
Autoconf chokes on spaces, but we may receive a full path from Cabal
containing spaces, so if that happens we just ignore Cabal's suggestion.
Terrible hack, but what can you do?  We can't just bail because Windows
users might have their compiler under "Program Files", so we'll just
print a warning and hope for the best.