Make foreign imports of realpath and utimensat safe
Although these functions never return EINTR, signals may still affect
these calls.  In particular, realpath on Mac OS X contains a
retry-if-interrupted loop on statfs64.  If the file system is slow
(e.g. SSHFS), the call will hang due to the barrage of periodic alarm
signals from the GHC runtime.  Marking the foreign import as safe
appears to silence the alarm signals, although it is not guaranteed [1].
This is the best we can do until [2] gets fixed.

We also do this for utimensat as a precautionary measure.  Given that
these are system calls that manipulate the file system, the performance
overhead ought to be negligible.

This should alleviate #35 for now.

[1] https://mail.haskell.org/pipermail/ghc-devs/2015-September/009770.html
[2] https://ghc.haskell.org/trac/ghc/ticket/10840