Fix trailing path sep behavior of canonicalizePath and makeAbsolute
It used to be that `canonicalizePath "."` returned a path without a
trailing slash, but now it does.  This is due to the use of `normalise`
in `makeAbsolute`, which adds extra slashes whenever `.` gets stripped
out at the end.  This commit restores the original behavior for
`canonicalizePath`, and also affects `makeAbsolute` in a similar manner.

Fixes #42.