canonicalizePath: Drop trailing slashes
After discussion with Duncan Coutts, it was found that the trailing
slash-preserving behavior was actually a bug on Windows.  This means
there is really no reason for the current, somewhat quirky behavior of
preserving trailing slashes.  However, it has been a while since the
change was made, so it would be safer to introduce this as a major
version bump.

The internal prependCurrentDirectory function has been reworked slightly
with regards to the behavior on empty paths, but this not have any
visible effect on the public API since they always end up normalizing
the result of prependCurrentDirectory in some way or another.

Fixes #63.