canonicalizePath: Deref file symlinks even if not last segment
Due to the use of splitPath instead of splitDirectories, this means that
if a symbolic link segment in the middle of the path points to a file,
doesPathExist will report that it does not exist, thus the symbolic link
will not be dereferenced.  The behavior is now changed to dereference as
much as possible.