Make internal modules visible
Some of them were already available (as required by tests), but they had
their documentation hidden.  Hackage insists on showing the module names
anyway.

There's not really anything to hide here: the only reason they are
internal is because they have no API stability guarantees, so may as
well just make everything public.  The only exception is Config because
there is nothing useful in there and the presence of exeExtension may
mislead users into thinking that it's not available in the public API.

There is a blank line in C_utimensat and Posix.  They are needed because
hsc's {-# LINE ... #-} pragmas break the Haddock documentation.