Remove withFileStatus to avoid long path restrictions
The implementations have been refactored to use a "Metadata"-based
interface, inspired by the Rust standard library.  It's essentially a
portable version of 'stat'.  So far, this interface will remain
internal, but it might become public someday.

The permissions-related functions still need to be cleaned up, however.
They are one of the last remaining places that do not support long paths
on Windows.