Revert long path support for setCurrentDirectory
SetCurrentDirectoryW does not accept long paths even with the \\?\
prefix, despite what the MSDN says.  It's probably best to revert the
change to reduce the risk of breakage.

Still, getCurrentDirectory gets to keep the fromExtendedLengthPath
postprocessing, in case the user sets to a path with \\?\.  The presence
of \\?\ in paths can cause surprises so it's best to strip it.

https://ghc.haskell.org/trac/ghc/ticket/13373#comment:6