Refactor permissions implementation
Permissions on Windows have been simplified to use the underlying
Windows API calls rather than the pseudo-POSIX interface (which chokes
on long paths).

The metadata API has been extended slightly to allow handling of raw
permission bits (Mode).  Unfortunately, we can't take full advantage of
it because the existing permissions API is really quirky and
inconsistent and we have to maintain compatibility.  It's also unclear
what the ideal mapping between "rwes" and "rwxrwxrwx" should be.

No changes in the semantics of the permissions have been introduced.