Make doesDirectoryExist and doesPathExist reject empty paths
This restores the old behavior prior to 1.3.1.1.

The internals have been refactored to use "simplify" instead of
"normalise", the latter of which is a bit buggy on certain paths,
especially on Windows.  The Windows implementation of
prependCurrentDirectory has been rewritten to use GetFullPathName, which
knows how to handle drive-relative paths better.

The new emptyToCurDir function helps explicitly mark places where empty
paths are accepted.  So far, it is limited to canonicalizePath,
makeAbsolute, the permission functions, and the timestamp functions.

Fixes #84.