Drop support for older base, filepath, unix, and Win32
This is done in preparation for AFPP support and to reduce legacy code
paths.

On Windows, stack runs into dependency loops on LTSes earlier than 12
due to the need for Cabal-2.2 imposed by filepath. Consequently, the
earlier base that can be tested on Windows is 4.11.0, so the lower bound
on base has been bumped up to 4.11.0, and so time is bumped to 1.8.0 to
align with base.